requirements.txt
google-genai
mcp
fastapi
uvicorn
python-dotenv
httpx
